[mod/#230] AddSportsInfoMapper 분리
